change source label for excerpts
Users can enable source of the featured image to show in post excerpts. This used the pre-text defined in the Overlay section. This could cause confusion if it says "source", because it could be interpreted as if this is the source of the text, not only the image.

This pull request changes the default pre-text for featured image sources below excerpts into "Featured image" and adds the `isc_featured_image_source_pre_text` filter to allow users to change it. -m Fixes