extend caption layout options
Thie pull request allows to extend the caption layouts.

It originates from the need to have a caption layout that doesnt jump as reported in #145 without breaking the existing version.

The PR moves the Overlay Position option into the default layout option so that it is clear they are related.