Move global list shortcode attributes into options
The Global List shortcode used some attributes to control which images it showed and how many per page. These were now moved into options in the settings. If the `included` or `per_page` attributes are set in the shortcode, they will override the global settings.

The appropriate explaination and examples in the settings were also removed.