Remove image-post meta data when a post is updated
Images that were removed from a post content or as a featured image still showed up as attached to that post in the global list. This PR removes posts from the `isc_image_posts` meta data attached to images, so that it can be rewritten the next time the post is visited in the frontend.