Show source options in media library overlay
We disabled the source options in the media library on post edit pages to prevent a conflict between these options and the source options in image blocks. This lead to users not being able to edit image sources for featured images, since there are no block options for them.

This commit shows the media library options now when the option to enable block support is disabled. This prevents both fields from showing up at the same time and allows users to select what works best for them.