Enable overlay positions for background images.
The overlay position option didn’t work with background images defined in inline CSS or the style tag since the script did always look for an image. It will now fall back to the container size to position the overlay according to the options. However, the container with the image still needs the `isc-source` class set manually. Otherwise, I’d fear that setting it automatically could break styles on many elements.