Introduced ISC\Analyze_HTML class
The new ISC\Analyze_HTML class contains functions solely for extracting information from HTML.

The class is used now already when looking for image IDs and src to place captions in the content.

Improved the description of matches in the regular expression that looks for images in the content and rewritten some code to make it easier to manipulate.