Unused images
Combine the previous "unused images" query with the query for empty source information and display them in one table, since they have a similar purpose and the Unused Images label for one of them was misleading.

Show the number and file size of really unused images in the Unused Images table.