Set custom_text as explicit default standard author
There was an inconsistency with the standard source. When ISC was freshly installed, none of the standard source options was visibly selected, but in the frontend, the custom text showed up. As soon as the options were saved – unchanged – the author name was selected as an option and also showing in the frontend.

This commit now defined the custom text as the default standard source also with unsaved settings. This seems the best choice and should not break any behavior for existing users.