Add a note to the media library modal when block options are enabled
When block options are enabled, the image source fields are displayed in the image block options instead of the media library that opens when adding or changing an image in the post editor. Some users were confused about this.

This commit adds a note about where to find the image source fields when they are missing from the media library modal.