Improve unused images check
On an older site, the number of unused images and file size was not calculated correctly because the attachment meta data seems to have changed over time. It seems to not have contained `filesize` in the past, but many of our checks depend on that.

This commit makes counting independent of `filesize`. It also counts `original_image` individually, which is only set in some cases.