Custom maybe_unserialize() function
With even WordPress core functions to unserialize options being unsafe if going through unknown data, I introduced a new function that is a copy of the core `maybe_unserialize()` with an additional flag to not allow classes.

Replaced the only usage of core’s `maybe_unserialize` with the new helper function.