Change overlay settings
Enabling the methods of how to display an image source in a different box than the options to fine-tune them felt a bit confusion for a while now. So I moved the enable box for Overlays into the Overlay settings sections and hide all detailed options until the option is enabled.

This is just the first of many commits to implement this for all options.