Reorder overlay content function
The function to add overlays to the content was processed in an order that didn’t allow custom overlays for non-standard `img` tags to be considered in the index, which prevented them from showing in the Global List and Per-page list. I reordered it and the images showed up right away.