Filter Media Library by image source
One now finds a new filter field in the list view of the media library to show only images with or without sources.

The filter logic in general can be extended by the modules.

I also added a link from the list of missing image sources to this new view for easier access.

If Pro is installed, one can then use the bulk edit feature to quickly fill in missing sources.