Prevent content from jumping on caption load
When rendering captions on pages that need longer to load or are uncached in the browser, the caption jumps when being positioned above the image

This solution is very simple. It hides the caption until it is loaded into position with JavaScript. Since the wrapper `isc-source` is included, which is always involved in rendering, this should never cause hidden sources, unless the user has some custom code.