Added `svg` to the list of image file extensions
I only added image file extensions to the list when the file format was naturally supported by WordPress. SVGs are not yet allowed as an upload into the media library. Though when a users allows it, they are handled properly and I didn’t find a reason not to whitelist them in general. This is only relevant for some features. They were already working for image sources when used in an `img` tag.