Prevent fatal error on comment edit pages
According to the WordPress documentation, the `add_meta_boxes` hook accepts only a `WP_Post` object. That is not the case and it could also accept others, like `WP_Comment` which would then lead to fatal errors.

I replaced the hook with a more specific one that made also the check for the post type redundand.