Fix lookup of scaled and rotated images
Images that were scaled or rotated on upload have a different filename in `wp_posts.guid` and `wp_postmeta`’s `_wp_attached_file`. Though we previously looked for the same file URL in both.

Now, the URL that `attachment_url_to_postid()` is looking for does not have `scaled` or `rotated` removed.